+++ src/components/statistics-selector/statistics-selector.component.js
@@ -17,6 +17,7 @@ module.exports = {
                 { id: types.FOUR_TWENTY, name: "420" },
                 { id: types.KILL_BULL, name: "Kill Bull" },
                 { id: types.GOTCHA, name: "Gotcha" },
+                { id: types.JDC_PRACTICE, name: "JDC Practice" },
             ]
         }
     },