+++ src/pages/leg-result/components/leg-result/leg-result.component.js
@@ -53,14 +53,14 @@ module.exports = {
         var match = this.input.match;
         var visits = leg.visits;
 
-        var chartMaxValue = match.match_type.id == types.X01 || match.match_type.id == types.X01HANDICAP ? leg.starting_score : 0;
+        var chartMaxValue = match.match_type.id === types.X01 || match.match_type.id === types.X01HANDICAP ? leg.starting_score : 0;
         var labels = [];
         var values = { }
-        for (var i = 0; i < legPlayers.length; i++) {
-            var player = legPlayers[i];
+        for (let i = 0; i < legPlayers.length; i++) {
+            const player = legPlayers[i];
 
-            var score = leg.starting_score + player.handicap
-            if (match.match_type.id == types.DARTS_AT_X) {
+            let score = leg.starting_score + player.handicap
+            if (match.match_type.id === types.DARTS_AT_X || match.match_type.id === types.GOTCHA) {
                 score = 0;
             }
             values[player.player_id] = [ score ];
@@ -68,29 +68,36 @@ module.exports = {
                 chartMaxValue = score;
             }
         }
-        var round = 0;
+        let round = 0;
         labels.push(round);
 
-        for (var i = 0; i < visits.length; i++) {
-            var visit = visits[i];
+        for (let i = 0; i < visits.length; i++) {
+            const visit = visits[i];
             if (i % legPlayers.length === 0) {
                 round++;
                 labels.push(round);
             }
-            var current = values[visit.player_id][values[visit.player_id].length - 1];
+            let current = values[visit.player_id][values[visit.player_id].length - 1];
             if (visit.is_bust) {
                 values[visit.player_id].push(current);
             }
             else {
-                if (match.match_type.id == types.SHOOTOUT || match.match_type.id == types.DARTS_AT_X || match.match_type.id === types.CRICKET ||
-                    match.match_type.id == types.AROUND_THE_CLOCK || match.match_type.id == types.AROUND_THE_WORLD || match.match_type.id == types.SHANGHAI ||
-                    match.match_type.id == types.TIC_TAC_TOE || match.match_type.id == types.BERMUDA_TRIANGLE) {
+                if (match.match_type.id === types.SHOOTOUT || match.match_type.id === types.DARTS_AT_X || match.match_type.id === types.CRICKET ||
+                    match.match_type.id === types.AROUND_THE_CLOCK || match.match_type.id === types.AROUND_THE_WORLD || match.match_type.id === types.SHANGHAI ||
+                    match.match_type.id === types.TIC_TAC_TOE || match.match_type.id === types.BERMUDA_TRIANGLE || match.match_type.id === types.JDC_PRACTICE) {
                     current = current + visit.score;
                     values[visit.player_id].push(current);
                     if (current > chartMaxValue) {
                         chartMaxValue = current;
                     }
-                } else if (match.match_type.id == types.KILL_BULL) {
+                } else if (match.match_type.id === types.GOTCHA) {
+                    if (visit.score === 0) {
+                        current = 0
+                    } else {
+                        current += visit.score;
+                    }
+                    values[visit.player_id].push(current);
+                } else if (match.match_type.id === types.KILL_BULL) {
                     if (visit.score === 0) {
                         current = leg.starting_score;
                     } else {