+++ src/pages/leg/components/button-input-table/button-input-table.component.js
@@ -6,7 +6,8 @@ module.exports = {
             compactMode: false,
             allButtonsMode: false,
             players: input.players,
-            current: this.findCurrent(input.players)
+            current: this.findCurrent(input.players),
+            currentDart: 0
         }
     },
 
@@ -15,10 +16,12 @@ module.exports = {
     },
 
     onButtonPress(score, multiplier, isUndo) {
+        this.state.currentDart = isUndo ? this.state.currentDart - 1 : this.state.currentDart + 1;
         this.emit('score-button-pressed', score, multiplier, isUndo);
     },
 
     onScoreUpdate(data) {
+        this.state.currentDart = 0;
         this.state.players = data.players;
         this.state.current = this.findCurrent(this.state.players);
         this.setStateDirty("players");